refactored NoteInformationView
